feat: add ChartAI data types and implement ChartAI server chart
- Introduced ChartAIData and MarketData interfaces for structured data handling.
- Added ChartAIServerChart component to visualize Bollinger Bands using RNKLineView.
- Updated HomeScreen to include navigation to ChartAI demo.
- Enhanced KLineView and related classes to support new Bollinger Bands label styles.
- Modified index.d.ts and index.js to accommodate new Bollinger Bands configuration options.
- Updated iOS classes to handle new Bollinger Bands drawing logic and styles.

+++ ios/Classes/HTKLineContainerView.swift
@@ -91,7 +91,7 @@ class HTKLineContainerView: UIView {
                     : ((CGFloat(safeVisibleStart) + 0.5) * oldItemWidth - oldOffset)
 
                 let previousSelectedIndex = self.klineView.selectedIndex
-                let selectedId: Int? = {
+                let selectedId: CGFloat? = {
                     if previousSelectedIndex < 0 || previousSelectedIndex >= previousModels.count {
                         return nil
                     }
